Finish basic pokemon generation system
Finished the pokemon basic generator, it takes
the species id and level and returns a pokemon
with stats similar to what you'd find in an
official game. They even got natures that affect
their stats!

I have yet to implement shiny pokemon and expand
the pokemon list for the next generations.